Convert libsql context server to git context server
- Updates configuration to handle git repository paths
- Adds environment variables for debugging
- Integrates with uvx and mcp-server-git command
- Adds logging support through env_logger
