+++ .gitignore
@@ -1,2 +1,3 @@
+.DS_Store
 /target
 extension.wasm
