+++ Cargo.toml
@@ -1,13 +1,15 @@
 [package]
-name = "libsql-context-server"
+name = "zed_context_server"
 version = "0.1.1"
 edition = "2021"
-publish = false
 
 [lib]
-path = "src/libsql-context-server.rs"
+path = "src/zed_context_server.rs"
 crate-type = ["cdylib"]
 
 [dependencies]
 serde = "1.0"
 zed_extension_api = "0.2.0"
+serde_json = "1.0"
+log = "0.4"
+env_logger = "0.9"
