+++ extension.toml
@@ -1,10 +1,14 @@
-id = "libsql-context-server"
-name = "Libsql Context Server"
-description = "Model Context Protocol Server for Libsql"
+id = "zed-context-server"
+name = "Zed Context Server"
 version = "0.1.1"
 schema_version = 1
-authors = ["Nicholas Quezada <nicholas.quezada@outlook.com>"]
-repository = "https://github.com/nicholasq/libsql-context-server"
+authors = ["Your Name <your.email@example.com>"]
+description = "A context server extension for Zed"
+repository = "https://github.com/yourusername/zed-context-server"
 
-[context_servers.libsql-context-server]
-name = "Libsql Context Server"
+[commands.git_info]
+title = "Git Info"
+description = "Get information about the current git repository"
+
+[context_servers.zed-context-server]
+name = "A Zed Context Server"
